backend/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
mysqlclient==2.2.0
Pygments==2.16.1
python-dotenv==1.0.0
requests==2.31.0
rich==13.5.2
soupsieve==2.5
urllib3==2.0.4
